Break out constants, helpers and serializers
